% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IScore.R
\name{create_mice_imputations}
\alias{create_mice_imputations}
\title{Very internal function for getting mice methods}
\usage{
create_mice_imputations(methods, maxit = 5)
}
\description{
Very internal function for getting mice methods
}
\examples{
methods <- c("pmm", "cart", "sample", "norm.nob", "DRF")
imputation_funcs <- create_mice_imputations(methods)

}
\keyword{internal}
