% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy_dist.R
\name{energy_dist}
\alias{energy_dist}
\title{Energy distance}
\usage{
energy_dist(X, X_imp)
}
\arguments{
\item{X}{a complete original dataset}

\item{X_imp}{an imputed dataset}
}
\description{
This function is a wrapper for energy distance calculating. For more details
see \link[energy]{eqdist.e}.
}
\examples{
X <- matrix(rnorm(1000), nrow = 100)
X_imp <- matrix(rnorm(1000), nrow = 100)
energy_dist(X, X_imp)

}
