% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IScore_cat.R
\name{factor_to_onehot}
\alias{factor_to_onehot}
\title{One hot encoding

A supplementarty function for one-hot encoding}
\usage{
factor_to_onehot(dat)
}
\arguments{
\item{dat}{a data containinig some factor but numeric columns.}
}
\description{
One hot encoding

A supplementarty function for one-hot encoding
}
