% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miceDRF.R
\name{impute_mice_drf}
\alias{impute_mice_drf}
\title{mice DRF (Distributional Random Forest)}
\usage{
impute_mice_drf(missdf, printFlag = FALSE, ...)
}
\arguments{
\item{missdf}{incomplete dataset with missing values denoted as NA's}

\item{printFlag}{logical, indicating whether silent computations should be
performed. Default to FALSE.}

\item{...}{used for compatibility with mice package.}
}
\value{
completed dataset
}
\description{
This function performs imputation using MICE and Distributional Random Forest
}
\examples{
X <- matrix(rnorm(1000), nrow = 100)
X[c(runif(700), rep(1, 300)) < 0.3] <- NA
impute_mice_drf(X)

}
\references{
This method is described in detail in:

Näf, J., Scornet, E., & Josse, J. (2024).
What is a good imputation under MAR missingness?. arXiv.
\url{https://arxiv.org/abs/2403.19196}

It's based on:

Cevid, D., Michel, L., Näf, J., Meinshausen, N., and B¨ uhlmann, P. (2022).
Distributional random forests: Heterogeneity adjustment and multivariate
distributional regression. Journal of Machine Learning Research, 23(333):1–79.
}
