% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IScore.R
\name{Iscore}
\alias{Iscore}
\title{Calculates score for a single imputation function}
\usage{
Iscore(X, X_imp = NULL, N = 50, imputation_func, max_length = NULL)
}
\description{
Calculates score for a single imputation function
}
