% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IScore_beta_v2.R
\name{Iscore_beta_v2}
\alias{Iscore_beta_v2}
\title{Calculates score for one imputation function}
\usage{
Iscore_beta_v2(
  X,
  X_imp,
  multiple = TRUE,
  N = 50,
  imputation_func,
  max_length = NULL,
  skip_if_needed = TRUE,
  n_patterns = 10,
  scale = FALSE
)
}
\arguments{
\item{n_patterns}{numerical indicating how many patterns should be sampled
to calculate the score for j'th column. \code{n_patterns} can be NULL meaning
that all the patterns will be used (it might be slow). Default to 10.}
}
\value{
a numerical value denoting weighted Imputation Score obtained for
provided imputation function and a table with scores and weights calculated
for particular columns.
}
\description{
Calculates score for one imputation function
}
\examples{
set.seed(111)
X <- matrix(rnorm(8000), nrow = 1000)
X[runif(8000) < 0.4] <- NA
X_imp <- impute_mice_drf(X)

miceDRF::Iscore_beta_v2(X, X_imp, N = 10, imputation_func = impute_mice_drf)

}
