% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IScore_cat.R
\name{Iscore_cat}
\alias{Iscore_cat}
\title{Calculates score for a single imputation function}
\usage{
Iscore_cat(
  X,
  X_imp,
  imputation_func,
  onehot = FALSE,
  mask = NULL,
  multiple = TRUE,
  N = 50,
  max_length = NULL,
  skip_if_needed = TRUE
)
}
\arguments{
\item{X}{data containing missing values denoted with NA's}

\item{X_imp}{imputed dataset.}

\item{imputation_func}{a function that imputes data}

\item{onehot}{a logical meaning the setup of categorical variables
imputation. I=f \code{TRUE} it means one-hot encoding for imputation.
Otherwise factors are used}

\item{mask}{an integer vector ocntaining repeated id's of factor variables.
Used only when \code{X_imp} and \code{X} are provided in one-hot encoding. It
should have the same length as the number of columns in proivided data frames.
It should contain 0's when the column does not describe vategory (for instance
numerical columns) and id's for categorical columns. (See details).}

\item{multiple}{a logical indicating whether provided imputation method is a
multiple imputation approach (i.e. it generates different values to impute
for each call). Default to TRUE. Note that if multiple equals to FALSE, N is
automatically set to 1.}

\item{N}{a numeric value. Number of samples from imputation distribution H.
Default to 50.}

\item{max_length}{Maximum number of variables \eqn{X_j} to consider, can
speed up the code. Default to \code{NULL} meaning that all the columns will
be taken under consideration.}

\item{skip_if_needed}{logical, indicating whether some observations should be
skipped to obtain complete columns for scoring. If FALSE, NA will be returned
for column with no observed variable for training.}
}
\value{
a numerical value denoting weighted Imputation Score obtained for
provided imputation function and a table with scores and weights calculated
for particular columns.
}
\description{
Calculates score for a single imputation function
}
\details{
If you are using the data already stored in one-hot variables you should provide
a \code{mask} variable. For example a mask variable for 6D data can look like
\code{c(1, 1, 0, 0, 0, 0)} which means that first two columns of provided data
store one one-hot encoded categorical variable.
}
\examples{
set.seed(123)
X <- matrix(rnorm(500), nrow = 100)
X <- cbind(X, factor(sample(1:5, 100, replace = TRUE), levels = 1:5))
X[runif(600) < 0.2] <- NA
X <- cbind(X, factor(sample(1:2, 100, replace = TRUE), levels = 1:2))
X <- as.data.frame(X)
X[["V6"]] <- factor(X[["V6"]], levels = 1:5)
X[["V7"]] <- factor(X[["V7"]], levels = 1:2)
X[["V8"]] <- rnorm(100)
imputation_func <- miceDRF:::create_mice_imputation("cart")
X_imp <- imputation_func(X)

Iscore_cat(X, X_imp, imputation_func, TRUE)

}
