% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IScore.R
\name{Iscores_new}
\alias{Iscores_new}
\title{IScore}
\usage{
Iscores_new(
  X,
  N = 50,
  imputation_funcs = NULL,
  imputations = NULL,
  max_length = NULL
)
}
\arguments{
\item{X}{data containing missing values denoted with NAs}

\item{N}{a numeric value. Number of samples from imputation distribution H.
Default to 50.}

\item{imputation_funcs}{A list of functions, whereby each
`imputation_funcs[[method]]` is a function that takes the arguments \code{X}
and \code{m} and imputes \code{X} \code{m} times using method:
\code{imputations = imputation_funcs[[method]](X,m)}. Default to NULL meaning
... TODO::}

\item{imputations}{Either \code{NULL} or a list of imputations for the
methods considered, each imputed \code{X} saved as
\code{imputations[[method]]}, whereby method is a string. Default to
\code{NULL}.}

\item{max_length}{Maximum number of variables \eqn{X_j} to consider, can
speed up the code}
}
\value{
a list of scores calculated for every imputation function from
\code{imputation_funcs}. Each result consists of two elements:

- a weighted score,
- a table with scores calculated for particular columns.
}
\description{
This function... TODO: details
}
\examples{
X <- matrix(rnorm(1000), nrow = 100)
X[c(runif(700), rep(1, 300)) < 0.3] <- NA
methods <- c("pmm", "cart", "sample", "norm.nob")
imputation_funcs <- create_mice_imputations(methods)
Iscores_new(X, N = 50, imputation_funcs)

}
