% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IScore.R
\name{Iscores_new_perimp}
\alias{Iscores_new_perimp}
\title{Calculates score for a single imputation function}
\usage{
Iscores_new_perimp(X, X_imp, N = 50, imputation_func, max_length = NULL)
}
\description{
Calculates score for a single imputation function
}
\keyword{internal}
