% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IScore_cat.R
\name{onehot_to_factor}
\alias{onehot_to_factor}
\title{One hot encoding}
\usage{
onehot_to_factor(onehot_dat)
}
\arguments{
\item{onehot_dat}{a data coded with \code{factor_to_onehot} function.}
}
\description{
A supplementarty function for one-hot encoding
}
