% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IScore.R
\name{create_mice_imputation}
\alias{create_mice_imputation}
\title{Very internal function for getting mice methods}
\usage{
create_mice_imputation(method)
}
\description{
Very internal function for getting mice methods
}
\examples{
methods <- "pmm"
imputation_funcs <- create_mice_imputations(methods)

}
\keyword{internal}
