% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_IScores.R
\name{create_mice_imputations}
\alias{create_mice_imputations}
\title{A function creating list of mice imputation functions}
\usage{
create_mice_imputations(methods)
}
\arguments{
\item{methods}{a character vector of names of mice methods}
}
\description{
A function creating list of mice imputation functions
}
\examples{
methods <- c("pmm", "cart", "sample", "norm.nob", "DRF")
reate_mice_imputations(methods)

}
