% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IScore.R
\name{Iscore}
\alias{Iscore}
\title{Calculates score for a single imputation function}
\usage{
Iscore(X, X_imp, multiple = TRUE, N = 50, imputation_func, max_length = NULL)
}
\arguments{
\item{X}{data containing missing values denoted with NA's}

\item{X_imp}{imputed dataset.}

\item{multiple}{a logical indicating whether provided imputation method is a
multiple imputation approach (i.e. it generates different values to impute
for each call). Default to TRUE. Note that if multiple equals to FALSE, N is
automatically set to 1.}

\item{N}{a numeric value. Number of samples from imputation distribution H.
Default to 50.}

\item{imputation_func}{a function that imputes data}

\item{max_length}{Maximum number of variables \eqn{X_j} to consider, can
speed up the code. Default to \code{NULL} meaning that all the columns will
be taken under consideration.}
}
\value{
a numerical value denoting weighted Imputation Score obtained for
provided imputation function and a table with scores and weights calculated
for particular columns.
}
\description{
Calculates score for a single imputation function
}
\examples{
X <- matrix(rnorm(1000), nrow = 100)
X[c(runif(700), rep(1, 300)) < 0.3] <- NA
imputation_func <- miceDRF:::create_mice_imputation("pmm")
X_imp <- imputation_func(X)

miceDRF::Iscore(X, X_imp, N = 50, imputation_func = imputation_func)

miceDRF::Iscore(X, X_imp, N = 50, imputation_func = imputation_func, multiple = FALSE)

# zero imputation
imputation_func <- function(X) {X[is.na(X)] <- 0; X}
X_imp <- imputation_func(X)

miceDRF::Iscore(X, X_imp, N = 50, imputation_func = imputation_func, multiple = FALSE)

}
