% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IScore.R
\name{Iscore}
\alias{Iscore}
\title{Calculates score for a single imputation function}
\usage{
Iscore(X, N = 50, imputation_func, max_length = NULL)
}
\description{
Calculates score for a single imputation function
}
\examples{
X <- matrix(rnorm(1000), nrow = 100)
X[c(runif(700), rep(1, 300)) < 0.3] <- NA
imputation_func <- miceDRF:::create_mice_imputation("pmm")
Iscore(X, N = 50, imputation_func = imputation_func)

}
