% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miceDRF.R
\name{mice.impute.DRF}
\alias{mice.impute.DRF}
\title{mice DRF (Distributional Random Forest)}
\usage{
mice.impute.DRF(
  y,
  ry,
  x,
  wy = NULL,
  min.node.size = 1,
  num.features = 10,
  num.trees = 10
)
}
\arguments{
\item{y}{words, words...}

\item{ry}{words, words...}

\item{x}{words, words...}

\item{wy}{words, words...}

\item{min.node.size}{words, words...}

\item{num.features}{words, words...}

\item{num.trees}{words, words...}
}
\description{
This function performs imputation using MICE and Distributional Random Forest
}
\references{
This method is described in detaik in Näf, J., Scornet, E., & Josse, J. (2024).
What is a good imputation under MAR missingness?. arXiv.
\url{https://arxiv.org/abs/2403.19196}

It's based on:

Cevid, D., Michel, L., N¨af, J., Meinshausen, N., and B¨ uhlmann, P. (2022).
Distributional random forests: Heterogeneity adjustment and multivariate
distributional regression. Journal of Machine Learning Research, 23(333):1–79.
}
